+++ src/compiler/transformer.ts
@@ -566,7 +566,7 @@ namespace ts {
     }
 
     export const nullTransformationContext: TransformationContext = {
-        factory,
+        factory: factory, // eslint-disable-line object-shorthand
         getCompilerOptions: () => ({}),
         getEmitResolver: notImplemented,
         getEmitHost: notImplemented,
